feat(StationEditor): insert location position when click on empty map place
