+++ src/main/scala/model/station/Location.scala
@@ -1,3 +1,5 @@
+package model.station
+
 import cats.Eq
 
 object Location:
