+++ src/main/scala/model/station/StationMap.scala
@@ -1,3 +1,5 @@
+package model.station
+
 import cats.implicits.catsSyntaxEq
 
 final case class StationMap(stations: List[Station]):
