+++ src/test/scala/LocationTest.scala
@@ -1,3 +1,4 @@
+import model.station.Location
 import org.scalatest.*
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
